+++ src/main/java/com/bugzero/rarego/global/exception/GlobalExceptionHandler.java
@@ -4,6 +4,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 
+import com.bugzero.rarego.global.response.ExceptionResponseDto;
+
 import lombok.extern.slf4j.Slf4j;
 
 @ControllerAdvice
